+++ src/blockCache/GlobalBlockCache.php
@@ -68,14 +68,15 @@ class GlobalBlockCache extends BlockCacheStrategy
     public function buildKey(BlockInterface $block): array
     {
         $key = [self::CACHE_TAG];
+        $identity = \Craft::$app->user->identity;
         if ($this->options->cachePerAuthenticated) {
-            $key[] = \Craft::$app->user ? 'auth' : 'noauth';
+            $key[] = $identity ? 'auth' : 'noauth';
         }
         if ($this->options->cachePerViewport) {
             $key[] = 'view-port-' . $this->getViewPort();
         }
-        if ($this->options->cachePerUser and $user = \Craft::$app->user) {
-            $key[] = 'user-id-' . $user->getIdentity()->id;
+        if ($this->options->cachePerUser and $identity) {
+            $key[] = 'user-id-' . $identity->id;
         }
         if ($this->options->cachePerSite) {
             $site = \Craft::$app->sites->getCurrentSite();